Initial commit
This commit adds the first working version of the monitoring and logging
of Twitch chat messages and events to a database.
Included is the full migration and schema. The building of the IRC
client and IRC message parsing. The querying of 7TV emote sets. And the
monitoring of whatever list of channels is given to update their live
status.
Not every aspect is finished in this commit. Missing is:
- Adding stream names to the database.
- Parsing of database messages to generate reports.
- Retrieval of BTTV and FrankerFaceZ emote sets. And;
- Clean up unused data structures and functions from previous
  uncommitted version testing.
- Fixing of a bug where global third party emotes aren't mentioned in
  database entries for users that have no account with them. I.E. using
  the global 7TV emote `glorp` on a channel that has no existing 7TV
  account will not mention the use of `glorp` in the `stream_message`
  entry.