Implements stream report generator and more.
This commit adds and changes a list of things.

Firstly, this commit adds the initial implementation for the stream report
generator. This app will take an id for a stream in the database, and
generate reports for it. The reports will be uploaded to pastebin, and
their links printed back. Because of the use of pastebun, a new config
option was added, `pastebin_api_key`. Which takes the value given at:
`https://pastebin.com/doc_api`.

Secondly, the `stream_message_emote` table has been removed. It
bloated the queries needed for parsing messages, which caused massive
delays in generation of the reports. The data has been moved to a new
column on the `stream_message` table, `stream_message.twitch_emote_usage`.
This column will contain a json string being a table with key `emote.id`
and value being how many times the emote was used in
`stream_message.contents`.

The remaining changes are a list of minor things.
- Cleans the app_config workspace, moving all the files out of the inner
  `config` directory.
- Removes the unnecessary `outdated_schema` file from the database migration.
- Removes an unused test in the irc message parser file.
- Adds a ping response to prevent connection timeouts.