Implements generating reports in file form
This commit adds a feature flag for the `database_report_generator` that
will spit out any reports for a stream in file form. This is for generating
any reports wanted without the need of uploading them to pastebin.
This also might be expanded upon later for generating private reports
that we might not want to upload anywhere.