Adds a README and reformats report files.
In order to improve the documentation of this project, this commit adds
a README file describing the setup of the config file.

In addition, this commit reforpats the report files to be a bit more
readable. And to automatically set the emote_message threshold instead
of hard coding the value.