Adds a script to run the report generator.
In order to make it easier to generate reports, this commit adds a
script that will take the stream ID and file flag. Then build and run
the `database_report_generator` app. This makes it easier to run the app
without having to build the app manually.

In addition, this commit adds some error handling and updates the README
with the remaining config setup.