Fixes misconfigured tags for raid parsing
In order to fix raid tracking, this commit fixes the values for tags
`RAID_VIEWER_COUNT` and `USER_ID` to be the correct, expected value.

In addition, this commit adds a test used for manually testing IRC
messages to see if the parser is adding them to the database correctly.