Adds unknown users for streamlabs donation events.
In order to account for modified donator names in streamlabs donations,
this commit updates the `donation_event` table to make the
`donator_twitch_user_id` nullable, with a new additional column for the
`unknown_user_id`. Due to the fact that users can insert any name into
the name field for streamlabs donations, just `twitch_user` cannot be
relied upon for setting the donator. To account for this, an additional
step is used to check if the name passed in is similar to any known
names using the jaro_winkler distance. If a match is found then the
twitch_user is used to identify a donation, otherwise a new
`unknown_user` is set for the donator.
When generating monthly reports, if there is no set `twitch_user` for a
donation. The `unknown_user` is used for the name instead.