Adds tracking for users changing their Twitch name
In order to account for users changing their names, this commit adds the
`twitch_user_name_change` table. This table will be created when a
mismatch is found between the existing name of a user in the database
and the one given back by Helix. Using Twitch's ID as an identifier.
In addition, this commit fixes a bug where the report generator app
wasn't compiling due to an update in the config's API.