Fixes report generator crashing with no donations
In order to prevent the report generator from crashing before there's
any donation data for the month, this commit makes the
donator_monthly_rankings report only get added to the list of reports to
print back when there's no error. This will prevent the entire program
from exiting when it fails.
In addition, this commit changes logging for the report generator to log
to stdout instead of to a file.