Implements a Kubernetes job for generating reports
In order to make the creation of reports as easy as possible, this
commit adds a Dockerfile and Kubernetes job config for running the
report generator. With this, a script was made to run, and tail the logs
of the job. As the output of the pastebins are still given through
stdout.

In addition, to allow for even easier running of the report generator.
This commit adds a feature where a name can be passed in, and reports
will be generated for the latest stream of that user.

Lastly, the `run_migrations` function in the `database_connection` crate
has been updated to include all of the tables added in previous commits
that were missing.