Refactors the codebase for better message parsing
In order to clean up the way message parsing is done, this commit
refactors parts of the codebase to account for deserializing IRC
messages directly instead of parsing their values at the location
they're wanted.
This is done through the `TwitchIrcTagValues` struct, which an
additional struct `TwitchIrcMessage` wraps around. TwitchIrcTagValues is
dedicated to deserialization, while TwitchIrcMessage is dedicated to
inferfacing with the values that were deserialized.
With these changes comes a complete overhaul of the parsing for each
message type. Message type is now calculated by TwitchIrcMessage.
The methods themselves have been greatly simplified by this wrapper
around the data as well.
Tests have also now been implemented for all message types that could be
received.

Additionally, in order to implement tests for the parser, a change in
how the database was passed around needed to be changed. Entity
extensions now require the database to be passed in the account for the
fact that tests run with a mock database. Additionally, all message
parsing methods require the database for the same reason.
Along with that, the entity extensions have been moved into their own
workspace crate with a `__test_hook` feature that will disable parts of
methods that don't need to be run in tests such as anything to do with
reqwest.
Along with this refactoring of the parser comes some other large changes
to account for testing and some minor improvements.

Along with this, to clean up the amount of errors, any errors that could
be combined, were combined. Those errors being the
`FailedToGetXAtY` and `FailedToParseXAtY` errors. These have been
replaced with `FailedToGetValue` and `FailedToParseValue` which take any
data as values on the enum such as location, value name, etc.

Lastly, some minor improvements to the codebase were done wherever they
were found to be done. These improvements include:
- Updating dependencies.
- renaming `get_most_recent_stream_for_user` to
  `get_active_stream_for_user` to be more clear as to what it's doing.
- Alias `irc::proto::Message` to IrcMessage wherever it was found.
- Properly implement Deserialize for SubTier. And;
- Disable the app animation when logging is set to go to stdout.