Fixes a bug with resubs and changes tests
In order to get all types of subscriptions as listed by the twitch API
here https://dev.twitch.tv/docs/chat/irc/#notice-tags
This commit adds checks for the `giftpaidupgrade` and
`anongiftpaidupgrade` message id's. These are for when someone is gifted
a subcription, then resubscribes. Sub tier and months are not given by
Twitch, but from past logs it's assumed this is only ever received when
it's the receiver wasn't subscribed before, and resubscribes at tier 1.
Because of this, the tags will set the tier and duration if the message
id is found to be `giftpaidupgrade` or `anongiftpaidupgrade`.