Remove anyhow from extensions crate and fixes bugs
In order to get more detailed and consistant information on errors from
the entity extensions crate, this commit removes the use of anyhow.
Replacing it instead with thiserror. The first use case of this is
implemented with the livestream updater, where if an error code of 503
(service unavailable) is received. The task will wait for 30 seconds
until starting up again. More like this will be implemented in the
future.

In addition, this commit fixes a bug with Twitch's changes to gift sub
IRC messages. They've updated singular gift subs to no longer give a
value for how many subs were gifted, resulting in errors. This has been
fixed by defaulting to 1 gift sub if the value is missing, and the
msg-id is one pertaining to gift subs.