Refactors for multithreaded support
In order to speed up the start of processing for each message, this
commit adds a system that will spawn a new task for each irc message
being parsed. This will both allow for the use of multiple threads, but
should also speed up the rate at which the irc stream is drained,
preventing potential ping timeouts.
Along with this, to ensure no messages are lost in the case of a
reconnect, this commit updates the reconnect method to process any
remaining irc messages in the client stream.

In order to account for this new way of processing messages, and to
clean up main a bit. This commit also refactors where the main and sub
processes live. Starting with spawning off any infinite tasks like
stream updates, and running a main task that checks for the results of
each message parsed.
Along with this, a new infinite task was made to receive the handles for
each message being parsed. This is to ensure that any errors that
occurred during parsing are logged.

With the addition of multithreading, in order to prevent a race
condition, this commit adds a check to inserting Twitch users. This
checks if the error returned is a `unique constraint violation`. If
another thread attempts to add a user at the same time, this is
returned, and one of the messages are lost as a result.