Fixes ping timeouts and reduces ping intervals
In order to resolve the case of a ping timeout, this commit changes how
the client reconnects. The previous way of doing it was not actually
reconnecting, and resulted in an infinite error loop. The new
implementation goes back to recreating a new instance of the irc config
obcject.
In addition, to shorted the time between reconnections, this commit
shortens the ping_timeout duration to 20 seconds from 30 and ping_time
interval to 60 seconds from 180.