Implements custom conditions in report generator
In order to increase the range of data that can be returned from report
template methods, this commit replaces the passing around of a stream ID
as a condition for queries with a custom set of conditions. This is
achieved through the AppQueryConditions which is build with any desired
data to constrain the data by. This allows for additional reports like
the `monthly_general_stats` report. Which requires a condition of both
the streamer and a range of time. Whereas the other existing reports
only need a stream id.

In addition, in order to be able to still generate donation reports for
private analysis, this commit adds an argument to the general_stats
report generator for generating donation data or not.