Implements use of EventSub API for stream statuses
In order to account for the switch to EventSub in the case of IRC
deprecation, this commit implements the full process of managing
EventSub subscription and message processing. For now the only use will
be for monitoring stream statuses. A full switch will be delayed as not
all features needed are possible without moderation permission for a
given channel.

In addition, a few other changes have been made to improve the codebase.
These are as follows:
- Moves any code out of mod and lib files in the main app. These include
  tracked_channels, twitch_irc, and processes.
- Removes stream status tracking from the TrackedChannels, leaving it
  only for storting channels that're wanted.
- Renames `channel/update_list` to `update_channel_live_status` in order
  to make the intent of the module be more clear.
- Resorts the check order for IRC message parsing to be in order of most
  to least frequent messages received.
- Adds a method for the EventSub stream status message parsing to the
  MessageParser. And implements tests to ensure it's parsing as
  expected.
- Adds `tungstenite` and `futures` as dependencies for handling of
  websocket connections.