Adds checks for active livestreams on app startup.
In order to track missing livestreams that started before the startup of
the app, this commit adds the `update_active_streams` method that will
check for any currently active streams, and add the missing ones to the
database.

In additon, this commit makes a few other minor changes. These changes
include:
- Updates SeaOrm to keep up with any new features and fixes.
- Moves `update_channel_status` and `app_animation` to the processes
  module to properly organize the codebase into its proper separation of
  concerns.
- Increases the subscription retry duration to 20 seconds to total 1
  minute of wait time for trying to resubscribe to events that failed.
- Moves any constants to do with the EventSub `keep alive` outside of
  the TwitchWebsocketConfig object.