Creates the `update_changed_names` app.
In order to update the database for anybody who changed their name on
Twitch's side, this commit creates the initial implementation for the
`update_changed_names` app. This app will check every name in the
database against their current name on Twitch, and update the entry if
there were any changes to the login or display name.