Updates stream message emote columns to json
In order to maintain clarity and to extend the use of queries for said
columns, this commit converts the `third_party_emotes_used` and
`twitch_emote_usage` from strings to json. Any instance of these columns
being used has been fixed to work, and tests were made to ensure values
are being returned as expected.