Implements tracking of gift sub recipients
In order to track which users are on the receiving end of a gift sub
donation, this commit adds the `gift_sub_recipient` bridge table for
storing which users received gift subs from a gift_sub donation event.
This table also contains the months that the user tied to it has been
subscribed for with this gift sub.

In order to use this table, a method has been added to the message
parser for checking if a gift sub is for a gift sub recipient, and adds
that data to the database. In addition, to track the required values for
this new table, additional tags have been added to be tracked to
retrieve the recipient twitch ID and months subscribed.