Fixes monthly general stats report not working
In order to fix the monthly general stats report, this commit updates
the condition passed in to the general stats report generator to be the
correct condition of the entire month instead of the previous stream. In
addition, this commit updates the `from_month` method under
AppQueryCondition to take a streamer twitch_user_id in order to separate
from other streamers in the database.