Implements a web query interface backend
In order to get a public web query up and running, this commit
implements the baseline interface backend to query for specific data
structures from the database. This implementation uses data transfer
objects to structure data from the database to return nested values for
each query.

In addition, in order to serialize the data to be sent, this commit
implements serde Serialize and Deserialize for the premade entities from
SeaORM.

Lastly, a method was added to the TwitchUserExtensions trait to obtain a
user from a condition without updating the database if the user didn't
exist.