Moves Docker and Kubernetes configuration files
In order to better organize the different Dockerfiles and Kubernetes
configuration files for the different apps for better clarity, this
commit moves the files into the respective project directories for each.
In addition, in order to clean up the Kubernetes configuration files for
apps that don't require Twitch access tokens, this commit makes those
fields optional in the AppConfig.