Fixes bugs and IRC client reconnection logic
In order to properly reconnect the IRC client in the event where
anything goes wrong (IE IO errors or anything wrong with reconnecting
the client itself), this commit changes the reconnection logic to
account for IO errors and no longer panic when the client couldn't
reconnect in the case of a DNS name resolution failure. In addition,
this commit removes the attempt to gather remaining messages from the
stream, as there was never anything to grab in the first place. This
added unncessary complexity to the code, and slowed down the
reconnection process.

Additionally, this commit updates the down migration for converting the
stream_message table's emote usage columns from json strings to json
types. This could fail in the event where the json value would serialize
into a string too large for the column. And a minor

Lastly, this commit changes the name changer Kubernetes CronJob to run
on Thursday nights in addition to Monday nights.