Fixes streamlabs donation parsing and Config bugs
In order to avoid missing streamlabs donations, this commit fixes a bug
that would skip user donations if the name area contained any spaces.
In addition, this commit abstracts the parsing of streamlabs donation
messages to make it consistent accross the project. As there is more
than one location that it needs parsing.

Lastly, this commit fixes the app config not appending channels to the
tracked channels list from the environment.