Logging to report generator and back-end ingress
In order to get a benchmark for the slowest reports to generate, this
commit adds logging of each benchmark call using tracing's increment
proc macro. In addition, to cut back on the noise when getting these
values, this commit adds a filter for SeaORM's logging to warn.

Additionally, in order to expose the REST API back-end to the internet,
this commit adds an ingress route for Traefik to route anything calling
`api.looank.com` to access the back-end service.