Initial implementation of React front-end
In order to get a more standardized front-end, this commit replaces the
existing initial implementation of the front-end with React through
Vite.
Adds a Dockerfile for the front-end for running in production.
Adds an ingress route to route traffic from the site to the front-end
app.
Updates the deployment and service for the front-end to use the correct
ports.