Adds websocket reconnect on n timeouts and fixes
In order to fix the websocket client dying without any message, this
commit adds a restart in the event that no messages are received from
Twitch after 5 timeouts. To achieve this the `check_for_stream_message`
method was updated to return True when a message was received, and
False if one was not.

In addition, this commit fixes a breaking bug where messages were not
being saved if they were no apart of a shared chat. This was a result of
erroring any time the `origin id` was missing. Turns out this value is
only sent when there is a shared chat active.

Lastly, this commit adds a fix to the name change checking script's rate
limiter. Tokens were not being reset after waiting for the rate limit
timer in the event of token exhaustion.