Implements support for shared chats and more.
In order to get proper support for shared chats, this commit adds the
source_id column to the `donation_event`, `subscription_event`, and
`user_timeout` tables as a unique key.
This will be set with the source id given by Twitch from messages
sent over shared chats.
In addition, this commit implements an ignore list for certain message
ids that will not be implemented. This will reduce the error log clutter
for messages that couldn't be parsed because they were never intended to
be parsed.