Adds additional features for the back-end app
In order to improve the scope of requesting data wanted, this commit
adds a way to get donation data in the back-end from a given channel in
the URL path.
In addition, in order to add a desired feature for the front-end to
guess a user based on their name, this commit implements an additional
feature on the users query to guess the login using a string that might
be within the desired name.

This commit also adds a few more minor changes. These are as follows:
- In order to improve debugging the back-end, adds tracing logging to
  the back-end.
- Properly implements axum's `IntoResponse` for the back-end AppError to
  improve clarity in errors and reduce the bloat of calling
  `into_response_error`.
- Adds Cors headers to allow calling back for data in the front-end