Converts emote storage to a join table and more.
In order to improve the way that emote data is stored within the
database, this commit removes storing third party and Twitch emotes
separately on the stream_message table. Instead, third party emotes have
been combined with the emote table. And instead of storing emote usage
data on the stream_message table, a join table is used for `emote_usage`
to join many messages to many emotes.