Add tailwind and remove shitty css
In order to remove the clown styling linkthedot made, this commit adds
tailwindcss to the project and replaces all of the external css files
with tailwind classes directly on components. It also makes the entire
website much more responsive. The results are now correctly shown in
cards for smaller screen devices and still in a table for larger
devices.