Improves project app names and moves chat tracker
In order to improve the clarity of apps after the scope creep of the
project, this commit moves the twitch chat tracker to its own app in the
workspace named `twitch_chat_tracker`.

In addition, this commit renames the report generator, back-end, and
front-end to be more clear and concise as to their functions.