Tests for emote dominant chat methods in reports
In order to improve the accuracy of reports generated by the report
generator with the new schema update, this commit implements tests for
the emote dominant chat related methods.
In addition, this commit fixes a bug in the `general_stats` report where
the `Total chats with < 70% emotes to words` data point was actually
listing the amount of chats with `> 70% emotes to words`.

Lastly, this commit adds a few minor changes. These changes are as
follows:
- Add missing tables to database table check.
- Adds helper methods in the report generator for tests to reduce
  bloat when creating required objects for testing.
- Adds another message id to the ignore list to reduce error log clutter
- Implements the use of the `month` argument in the report generator.