Implements scrubbing and processing of Spanix data
In order to expand the amount of data stored in the database, this
commit creates a module that will go through every known user in the
database and scrub every message for that user from Spanix. With another
feature flag, every message that's been stored for each user will be
process.