Rewrite chat ranking report and adds a column
In order to improve the speed and clarity of the chat ranking report
code, this commit rewrites it to be more expandable and readable for
future additions or edits.

In addition, this commit adds an additional column `%_of_all_words`
which shows how much of all words sent that stream were from the given
user on that row.