Adds monthly chat ranking flag for reports
In order to implement conditionally generating the monthly chat ranking
report, this commit adds an optional flag for the report generator that
will add it to the list of reports to generate.

Adding this made the previous method for generating the report list a
bit unwieldy. To fix that, this commit also moves the report generation
list to its own object in its own module. The generation for both
baseline reports (ones that will always be made no matter flags or
available data) and conditional reports (ones that may or may not be
generated based on existing data or flags).