Refactors the report generator
In order to account for a more expandable design that allows for
creating more report templates to be generated, this commit refactors
the report generator a bit to be centered around a more modular template
building design. Using a mixture of Tera and Tabled to build sections of
reports based on template files to reduce the amount of code needed to
be generated, and to separate how each template is created from the
data generation itself.

In addition, this commit adds subathon reports that can be generated by
passing in a start date along with a flag to indicate which report
template to generate.